requirements.txt

fastapi==0.111.0
uvicorn[standard]==0.30.1
faster-whisper==1.0.3
numpy>=1.26.0
sounddevice==0.4.6
webrtcvad==2.0.10
websocket-client==1.8.0
